Added meson build files
This adds support for meson build system.
For now, the build has only been tested on a linux host machine,
and does not properly implement the sound backend detection.
This patch does not remove any file, so that it is still possible
to use the autotools.
To fit in the meson logics of subprojects, some symbolic links
have been added, so that this might not work at all on windows.
Is the meson build system is finally adopted for julius, then
some directories should be moved as required.

Signed-off-by:Thierry Bultel <thierry.bultel@iot.bzh>